+++ srcCpp/RTIDDSImpl.h
@@ -6,6 +6,7 @@
  * Subject to Eclipse Public License v1.0; see LICENSE.md for details.
  */
 
+#include <stdexcept> // This header is part of the error handling library.
 #include <string>
 #include <algorithm>
 #include <map>
@@ -14,16 +15,21 @@
 #include "PerftestTransport.h"
 #include "RTIDDSLoggerDevice.h"
 
+
+#ifdef RTI_CUSTOM_TYPE
+#include "CustomType.h"
+#endif
+
 #define RTIPERFTEST_MAX_PEERS 1024
 
 /* Class for the DDS_DynamicDataMemberId of the type of RTI Perftest*/
 class DynamicDataMembersId
 {
-  private:
+private:
     std::map<std::string, int> membersId;
     DynamicDataMembersId();
 
-  public:
+public:
     ~DynamicDataMembersId();
     static DynamicDataMembersId &GetInstance();
     int at(std::string key);
@@ -32,7 +38,7 @@ class DynamicDataMembersId
 template <typename T>
 class RTIDDSImpl : public IMessaging
 {
-  public:
+public:
 
     RTIDDSImpl() :
         _transport(),
@@ -94,6 +100,12 @@ class RTIDDSImpl : public IMessaging
         _typename = T::TypeSupport::get_type_name();
 
         _pongSemaphore = NULL;
+
+        _qoSProfileNameMap[LATENCY_TOPIC_NAME] = std::string("LatencyQos");
+        _qoSProfileNameMap[ANNOUNCEMENT_TOPIC_NAME]
+                = std::string("AnnouncementQos");
+        _qoSProfileNameMap[THROUGHPUT_TOPIC_NAME]
+                = std::string("ThroughputQos");
     }
 
     ~RTIDDSImpl()
@@ -126,8 +138,9 @@ class RTIDDSImpl : public IMessaging
 
     DDSTopicDescription *CreateCft(const char *topic_name, DDSTopic *topic);
 
+    const std::string get_qos_profile_name(const char *topicName);
 
-  private:
+private:
 
     // Specific functions to configure the Security plugin
   #ifdef RTI_SECURE_PERFTEST
@@ -210,12 +223,13 @@ class RTIDDSImpl : public IMessaging
     RTIOsapiSemaphore *_pongSemaphore;
     RTIDDSLoggerDevice _loggerDevice;
 
-  public:
+    std::map<std::string, std::string> _qoSProfileNameMap;
+
+public:
 
     static int          _WaitsetEventCount;
     static unsigned int _WaitsetDelayUsec;
 };
-
 
 #endif // __RTIDDSIMPL_H__