+++ srcCpp/perftest_cpp.h
@@ -142,7 +142,11 @@ class perftest_cpp
   #endif
 
     // Number of bytes sent in messages besides user data
+  #ifdef RTI_CUSTOM_TYPE
+    static const int OVERHEAD_BYTES = 28 + 4; // For custom_type_size
+  #else
     static const int OVERHEAD_BYTES = 28;
+  #endif
     // Flag used to indicate message is used for initialization only
     static const int INITIALIZE_SIZE = 1234;
     // Flag used to indicate end of test