+++ srcCpp03/perftest_publisher.cxx
@@ -349,13 +349,13 @@ bool perftest_cpp::ParseConfig(int argc, char *argv[])
     /*
      * PERFTEST-108
      * We add this boolean value to check if we are explicity changing the
-     * number of iterations via command line paramenter. This will only be
+     * number of iterations via command-line paramenter. This will only be
      * used if this is a latency test to decrease or not the default number
      * of iterations.
      */
     bool numIterSet = false;
 
-    // Load command line parameters.
+    // Load command-line parameters.
     for (i = 0; i < argc; ++i) {
         if (IS_OPTION(argv[i], "-pub")) {
             _IsPub = true;
@@ -722,13 +722,20 @@ bool perftest_cpp::ParseConfig(int argc, char *argv[])
         }
         else if (IS_OPTION(argv[i], "-executionTime"))
         {
-            if ((i == (argc-1)) || *argv[++i] == '-')
-            {
+            if ((i == (argc-1)) || *argv[++i] == '-') {
                 std::cerr << "[Error] Missing <seconds> after -executionTime" << std::endl;
                 throw std::logic_error("[Error] Error parsing commands");
 
             }
+
             _executionTime = (unsigned int) strtol(argv[i], NULL, 10);
+
+            if (_executionTime <= 0) {
+                std::cerr << "[Error] -executionTime value must be a positive number "
+                          << "greater than 0"
+                          << std::endl;
+                return false;
+            }
 
         } else if (IS_OPTION(argv[i], "-cpu"))
         {